Added initial implementation and project description
Command line tool to extract email addresses associated to a specific
username from certain websites. This project was inspired by
github-email a shell script that communicates with GitHub's API service
to retrieve the email address associated to an account. This project
aims to offer the same functionality and expand to other popular
websites to add more reliability; notice that in comparison with the
original idea this version of the code will delete duplicate entries as
well as clear non-responsive addresses.