Added function to retrieve email from GitHub profile
The first attempt made by this program to retrieve an email address is
to send a HTTP GET request to GitHub's API service, however they have a
rate limit protection to prevent abusive usage. You can go around this
by sending the requests through a proxy or creating a token that will
allow you to send more requests but still with a limitation. This tool
assumes that you will not do any of these two things so the second
attempt will be to send a HTTP request to the GitHub profile page and
URL decode the email address if available.