Added function to retrieve the email from the events
The program will attempt to retrieve the email address from three
different sources, first using GitHub's public API service, if this
fails due to the rate limit protection the program will continue with a
request to the GitHub profile page which usually contains an URL encoded
version of the main email address, if this attempt fails the program
will continue sending a request to the GitHub activity page which is
part of the public API service. If the first attempt failed this third
option will surely fail, but I will keep it here in case that the API is
not rate limited but the address is not available in none of the first
two sources.