Added utils to collect emails before printing and struct
I rewrote the code into an struct to share the collected emails among
the other functions, this will facilitate the removal of the duplicated
addresses and some other goodies that I will add in the future. Notice
how we try to get the email from the API before the profile page even
when the API service is rate limited, this is because the data returned
by the API is more reliable. If we extract the email from the profile
page even when the API is still available the string might contain a
fake address.