Added rate limit detection for users API endpoint
There are two functions that rely on the availability of the API service
which is rate limited, one is always executed after the other, it makes
sense to stop the subsequent HTTP requests to the API if the first one
fails. In this case, if the IP address is trap in the rate limit block
the program will always use the profile page as the data source.