Modified to execute HTTP requests in concurrent mode
Using this program to retrieve the email address associated to certain
username and additionally the email address of the following accounts
takes around ~20 secs for an account with 40 friends. This commit
changes the behaviour of the code to use concurrency and takes the
execution time down to ~3.5 secs for the same test case.