Added option to retrieve the emails from the followers
Additional to the list of friends that an user account can follow via
GitHub there is also the option for other people to follow you, their
username gets added to a second page named Followers which contains a
maximum of 51 entries per page (as of August, 2016) divided in groups
of three. This options gets the email of the users listed in the first
page.