Added option to get the usernames instead of emails
This option will force the program to print a list of usernames found in
the following or follower pages for the specified page number, or the
first page by default. This is useful if you want to perform additional
operations over a list of usernames instead of getting the emails.